change WiimoteController
